Move settings view components to separate folders
