+++ desktop/packages/mullvad-vpn/src/renderer/components/views/settings/components/index.ts
@@ -1,0 +1,10 @@
+export * from './api-access-methods-list-item';
+export * from './app-info-list-item';
+export * from './daita-list-item';
+export * from './debug-list-item';
+export * from './multihop-list-item';
+export * from './quit-button';
+export * from './split-tunneling-list-item';
+export * from './support-list-item';
+export * from './user-interface-settings-list-item';
+export * from './vpn-settings-list-item';
