+++ desktop/packages/mullvad-vpn/src/renderer/components/views/settings/components/quit-button/index.ts
@@ -1,0 +1,1 @@
+export * from './QuitButton';
