+++ desktop/packages/mullvad-vpn/src/renderer/components/views/settings/components/split-tunneling-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from './SplitTunnelingListItem';
