+++ desktop/packages/mullvad-vpn/src/renderer/components/views/settings/components/support-list-item/SupportListItem.tsx
@@ -1,0 +1,19 @@
+import { messages } from '../../../../../../shared/gettext';
+import { Icon } from '../../../../../lib/components';
+import { ListItem } from '../../../../../lib/components/list-item';
+import { RoutePath } from '../../../../../lib/routes';
+import { NavigationListItem } from '../../../../NavigationListItem';
+
+export function SupportListItem() {
+  return (
+    <NavigationListItem to={RoutePath.support}>
+      <ListItem.Label>
+        {
+          // TRANSLATORS: Navigation button to the 'Support' view
+          messages.pgettext('settings-view', 'Support')
+        }
+      </ListItem.Label>
+      <Icon icon="chevron-right" />
+    </NavigationListItem>
+  );
+}
