+++ desktop/packages/mullvad-vpn/src/renderer/components/views/settings/components/user-interface-settings-list-item/UserInterfaceSettingsListItem.tsx
@@ -1,0 +1,19 @@
+import { messages } from '../../../../../../shared/gettext';
+import { Icon } from '../../../../../lib/components';
+import { ListItem } from '../../../../../lib/components/list-item';
+import { RoutePath } from '../../../../../lib/routes';
+import { NavigationListItem } from '../../../../NavigationListItem';
+
+export function UserInterfaceSettingsListItem() {
+  return (
+    <NavigationListItem to={RoutePath.userInterfaceSettings}>
+      <ListItem.Label>
+        {
+          // TRANSLATORS: Navigation button to the 'User interface settings' view
+          messages.pgettext('settings-view', 'User interface settings')
+        }
+      </ListItem.Label>
+      <Icon icon="chevron-right" />
+    </NavigationListItem>
+  );
+}
