+++ desktop/packages/mullvad-vpn/src/renderer/components/views/settings/components/vpn-settings-list-item/VpnSettingsListItem.tsx
@@ -1,0 +1,19 @@
+import { messages } from '../../../../../../shared/gettext';
+import { Icon } from '../../../../../lib/components';
+import { ListItem } from '../../../../../lib/components/list-item';
+import { RoutePath } from '../../../../../lib/routes';
+import { NavigationListItem } from '../../../../NavigationListItem';
+
+export function VpnSettingsListItem() {
+  return (
+    <NavigationListItem to={RoutePath.vpnSettings}>
+      <ListItem.Label>
+        {
+          // TRANSLATORS: Navigation button to the 'VPN settings' view
+          messages.pgettext('settings-view', 'VPN settings')
+        }
+      </ListItem.Label>
+      <Icon icon="chevron-right" />
+    </NavigationListItem>
+  );
+}
